Time: 18 ms (22.82%) | Memory: 7.7 MB (64.66%) - LeetSync

+++ 1013-fibonacci-number/fibonacci-number.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    int fib(int n) {
+    int fib(double n) {
         if (n == 0)
             return 0;
         if (n == 1)
